requirements.txt
pandas==2.2.3
openai==1.61.1
python-dotenv==1.0.1
paramiko==3.5.1
scikit-learn==1.6.1
seaborn==0.13.2
matplotlib==3.10.0
pytest==8.3.4
google-cloud-storage==3.0.0
gradio==5.15.0
sentence_transformers==3.4.1